Move gRPC-related code into oak::grpc sub-module
